[fix]: [CDS-78115]: change pdc infra validation  logic
