+++ 125-cd-nextgen/src/main/java/io/harness/cdng/creator/variables/DeploymentStageVariableCreator.java
@@ -42,6 +42,7 @@ import io.harness.cdng.visitor.YamlTypes;
 import io.harness.data.structure.CollectionUtils;
 import io.harness.data.structure.EmptyPredicate;
 import io.harness.encryption.Scope;
+import io.harness.evaluators.ProvisionerExpressionEvaluator;
 import io.harness.executions.steps.StepSpecTypeConstants;
 import io.harness.ng.core.environment.beans.Environment;
 import io.harness.ng.core.environment.mappers.EnvironmentMapper;
@@ -412,7 +413,8 @@ public class DeploymentStageVariableCreator extends AbstractStageVariableCreator
         infrastructureConfig.getInfrastructureDefinitionConfig().getSpec(), EnvironmentOutcome.builder().build(),
         ServiceStepOutcome.builder().build(), infrastructureEntity.getAccountId(),
         infrastructureEntity.getOrgIdentifier(), infrastructureEntity.getProjectIdentifier(),
-        infrastructureConfig.getInfrastructureDefinitionConfig().getTags());
+        infrastructureConfig.getInfrastructureDefinitionConfig().getTags(),
+        new ProvisionerExpressionEvaluator(Collections.emptyMap()));
 
     List<String> infraStepOutputExpressions =
         VariableCreatorHelper.getExpressionsInObject(infrastructureOutcome, OutputExpressionConstants.INFRA);
